+++ includes/plugins/one-click-demo-import/class-one-click-demo-import.php
@@ -6,7 +6,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    2.0.0
- * @version  2.0.2
+ * @version  2.4.0
  *
  * Contents:
  *
@@ -133,7 +133,7 @@ class Modern_One_Click_Demo_Import {
 		 * Info texts
 		 *
 		 * @since    2.0.0
-		 * @version  2.0.0
+		 * @version  2.4.0
 		 *
 		 * @param  string $text  Default intro text.
 		 */
@@ -158,7 +158,7 @@ class Modern_One_Click_Demo_Import {
 				$text .= '<div class="ocdi__demo-import-notice">';
 
 					$text .= '<h3>';
-					$text .= esc_html__( 'Install required plugins!', 'modern' );
+					$text .= esc_html__( 'Install demo required plugins!', 'modern' );
 					$text .= '</h3>';
 
 					$text .= '<p>';
@@ -167,6 +167,12 @@ class Modern_One_Click_Demo_Import {
 					$text .= ' <a href="https://github.com/webmandesign/demo-content/tree/master/modern/content#before-you-begin" title="' . esc_attr__( 'Read the information before you run the theme demo content import process.', 'modern' ) . '"><strong>';
 					$text .= esc_html__( 'View the list of required plugins &raquo;', 'modern' );
 					$text .= '</strong></a>';
+					$text .= '</p>';
+
+					$text .= '<p>';
+					$text .= '<em>';
+					$text .= esc_html__( '(Note that this set of plugins may differ from plugins recommended under Appearance &rarr; Install Plugins!)', 'modern' );
+					$text .= '</em>';
 					$text .= '</p>';
 
 				$text .= '</div>';