+++ includes/welcome/class-welcome.php
@@ -6,14 +6,15 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    2.0.0
- * @version  2.0.0
+ * @version  2.4.0
  *
  * Contents:
  *
- *  0) Init
- * 10) Renderer
- * 20) Admin menu
- * 30) Assets
+ *   0) Init
+ *  10) Renderer
+ *  20) Admin menu
+ *  30) Assets
+ * 100) Others
  */
 class Modern_Welcome {
 
@@ -195,6 +196,56 @@ class Modern_Welcome {
 					wp_enqueue_style( 'modern-welcome' );
 
 		} // /assets
+
+
+
+
+
+	/**
+	 * 100) Others
+	 */
+
+		/**
+		 * Info text: Rate the theme.
+		 *
+		 * @since    2.4.0
+		 * @version  2.4.0
+		 */
+		public static function get_info_like() {
+
+			// Output
+
+				return sprintf(
+					esc_html_x( 'If you %1$s like this theme, please rate it %2$s', '%1$s: heart icon, %2$s: star icons', 'modern' ),
+					'<span class="dashicons dashicons-heart" style="color: red; vertical-align: middle;"></span>',
+					'<a href="https://wordpress.org/support/theme/modern/reviews/#new-post" style="display: inline-block; color: goldenrod; text-decoration-style: wavy; vertical-align: middle;"><span class="dashicons dashicons-star-filled"></span><span class="dashicons dashicons-star-filled"></span><span class="dashicons dashicons-star-filled"></span><span class="dashicons dashicons-star-filled"></span><span class="dashicons dashicons-star-filled"></span></span></a>'
+				)
+				. '<br>'
+				. '<a href="http://webmandesign.eu/contact/?utm_source=modern">'
+				. esc_html__( 'And/or please consider a donation, thank you 🙏😊', 'modern' )
+				. '</a>';
+
+		} // /get_info_like
+
+
+
+		/**
+		 * Info text: Contact support.
+		 *
+		 * @since    2.4.0
+		 * @version  2.4.0
+		 */
+		public static function get_info_support() {
+
+			// Output
+
+				return
+					esc_html__( 'Have a suggestion for improvement or something is not working as it should?', 'modern' )
+					. ' <a href="https://support.webmandesign.eu/">'
+					. esc_html__( 'Contact support center &rarr;', 'modern' )
+					. '</a>';
+
+		} // /get_info_support