+++ template-parts/admin/welcome-footer.php
@@ -8,7 +8,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    2.0.0
- * @version  2.0.0
+ * @version  2.4.0
  */
 
 
@@ -18,5 +18,11 @@
 ?>
 
 </div> <!-- /.welcome-content -->
+
+<p>
+	<?php echo Modern_Welcome::get_info_support(); ?>
+	<br>
+	<?php echo Modern_Welcome::get_info_like(); ?>
+</p>
 
 <p><small><em><?php esc_html_e( 'You can disable this page in Appearance &raquo; Customize &raquo; Theme Options &raquo; Others.', 'modern' ); ?></em></small></p>