+++ template-parts/admin/welcome-promo.php
@@ -8,7 +8,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    2.0.0
- * @version  2.0.0
+ * @version  2.4.0
  */
 
 
@@ -26,18 +26,18 @@
 		(<a href="https://www.webmandesign.eu" target="_blank"><?php esc_html_e( 'Go to WebMan Design website &raquo;', 'modern' ); ?></a>)
 		<?php esc_html_e( 'Or perhaps you are considering a small donation?', 'modern' ); ?>
 		&rarr;
-		<a href="https://www.paypal.me/webmandesign/20" target="_blank"><em><?php esc_html_e( '"Hey Oliver, have a gallon of coffee on me :)"', 'modern' ); ?></em></a>
+		<a href="http://webmandesign.eu/contact/?utm_source=modern" target="_blank"><em><?php esc_html_e( '"Hey Oliver, have a gallon of coffee on me :)"', 'modern' ); ?></em></a>
 	</p>
 
 	<p>
 		<?php esc_html_e( 'You can also rate the theme at WordPress repository page.', 'modern' ); ?>
-		<a href="https://wordpress.org/support/theme/modern/reviews/?filter=5">
+		<a href="https://wordpress.org/support/theme/modern/reviews/#new-post">
 			<?php esc_html_e( "Let's go and rate the theme &#9733;&#9733;&#9733;&#9733;&#9733; :)", 'modern' ); ?>
 		</a>
 	</p>
 
 	<p>
-		<a href="https://www.paypal.me/webmandesign/20" target="_blank" class="button button-primary button-hero"><?php esc_html_e( 'Support theme development', 'modern' ); ?></a>
+		<a href="http://webmandesign.eu/contact/?utm_source=modern" target="_blank" class="button button-primary button-hero"><?php esc_html_e( 'Support theme development', 'modern' ); ?></a>
 	</p>
 
 	<p class="welcome-upgrade-thanks">