+++ template-parts/meta/entry-meta-element-comments.php
@@ -6,7 +6,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    2.0.0
- * @version  2.0.0
+ * @version  2.4.0
  */
 
 
@@ -15,7 +15,10 @@
 
 // Requirements check
 
-	if ( ! comments_open( get_the_ID() ) ) {
+	if (
+		post_password_required()
+		|| ! comments_open( get_the_ID() )
+	) {
 		return;
 	}
 
@@ -28,7 +31,7 @@
 ?>
 
 <span class="entry-meta-element comments-link">
-	<a href="<?php echo esc_url( get_permalink( get_the_ID() ) ); ?>#comments" title="<?php echo esc_attr( sprintf( esc_html_x( 'Comments: %s', '%s: number of comments.', 'modern' ), number_format_i18n( $comments_number ) ) ); ?>">
+	<a href="<?php the_permalink(); ?>#comments" title="<?php echo esc_attr( sprintf( esc_html_x( 'Comments: %s', '%s: number of comments.', 'modern' ), number_format_i18n( $comments_number ) ) ); ?>">
 		<span class="entry-meta-description">
 			<?php echo esc_html_x( 'Comments:', 'Post meta info description: comments count.', 'modern' ); ?>
 		</span>